+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 # CHANGELOG
 
+* Support HTTPS as health check protocol (@timoreimann)
+
 ## v0.1.26 (beta) - June 16th 2020
 
 * Update Kubernetes dependences to 1.18.3 (@waynr)